Add all the changes made and their tests
This commit is made due to the fact that I did not commit as I went
along whilst going through this challenge. This is why I am committing so
many files in one go. The code seems to work fine for now but there are
a few pending/ failing tests that I cannot fix for the time being.